Update patch set 2
Patch Set 2:

As you say, if the hypothetical prettytable 0.7.3 has this problem then the gate breaks.  This is intentional as we need to know if that is the case. [see below for more on that.]  This script is meant to address short-term fixes for upstream brokenness.  It performs the package installs so we can fix their brokenness before pip starts resolving the dependencies on our code.  It does assume that we're going to ignore (or overwrite if you're on CentOS) a system package that may be present is true, but pip would do that later anyway if the system package does not meet requirements.txt

The good thing about the permissions fixes is that Donald cleaned this up for us in pip 1.4.2 so once we can assume that pip 1.4 is not being used we can remove them.  And once we know the gate has >= 1.4.2 we can forget about worrying about a broken prettytable 0.7.3 or whatever there.

The py26 stuff does need to be handled, thanks.

Patch-set: 2