Update patch set 1
Patch Set 1:

Just say that if you want to 'daemonize' a non-daemon process (which is what is going on here) you need to ensure that bash evaluates the command-line args in a subshell.  That can be done in a couple of ways, commonly done in devstack by using multiple commands (include ';'), I/O redirection (include '<', '>' or '|') or a command block (enclose in '{}').

Patch-set: 1