Update patch set 2
Patch Set 2:

The indent check needs to be a bit smarter.  It needs state to account for here documents and possibly multi-line strings.  Multi-line commands account for a portion of what it reports now too, but those can be fixed without impact.

We can live with this for now as we need to fix up DevStack before we gate anyway.

Patch-set: 2