Update patch set 2
Patch Set 2: Looks good to me (core reviewer)

We do have here docs, and there are a few with indentation concerns:

 * lib/neutron_plugins/cisco - a python snippet for a version check
 * tools/create_userrc.sh  - help output.  
 * tools/xen/scripts/mkxva - usage output, error string, grub config stanzas

There is one in stack.sh (syslog) that has hard tabs that need to stay but they are correctly ignored.

The multi-line strings are mostly, if not all, sed commands that can^H^H^Hshould be indented properly.

Let's go ahead and put it in and tweak on it while we fix the problems that need fixing anyway.

Patch-set: 2
Label: Code-Review=+2