fixup_stuff.sh: don't break on prettytable perms fixup on RHEL
The permissions fixup hack assumes that prettytable-0.7.2 is
installed. Since python-prettytable 0.6.1 is already installed
from install_prereqs.sh (via files/rpm/general) and pip_install
does not set --upgrade, that assumption is inaccurate.

Apply the permissions fix to a more inclusive shell glob so
as to not fail when attempting to chmod a non-existant file.

Change-Id: I50d3b29ff1e2e0c5451700adaa4a634cf70209fc