Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

I just submitted https://review.openstack.org/#/c/51939/ to complete the steps for upper-layer projects to be included in DevStack without needing to modify the core scripts. This also allows projects not in the DevStack repo to easily be added without requiring patches to DevStack.

I've put an example of the conversion of the patch to add Savanna at https://github.com/dtroyer/devstack/tree/extras-savanna. The main change is to move the additions to stack.sh, unstack.sh and clean.sh into a new extras.d/70-savanna.sh script that will be called at the right time.

I want to start this conversion with the two projects that are pending addition to DevStack (Savanna and Marconi) and once those are working we will convert some of the existing projects to this format.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0