Update patch set 1
Patch Set 1:

The prettytable bit is causing problems on RHEL, I get that.  It also is supposed to only apply to 0.7.2 as that package is broken.

I'm going through all of the attempts to fix this and will figure out which one does what I want: fix 0.7.2 if present and leave everything else alone.  I specifically don't want to install 0.7.2 on top of the OS package on RHEL as the python config on RHEL is broken and installs pip-installed packages on top of the OS-installed package.

Patch-set: 1