Update patch set 1
Patch Set 1:

I've worked over my docs script into releaseable form and proposed it in https://review.openstack.org/52226.  It is a first cut but handles all of the shocco processing for devstack.org.  It should be enough for you to duplicate what I'm doing if you run it with no args.  It'll create a devstack-docs-* directory inside based on the github master.  The major thing lacking yet besides spit-n-polish is the ability to push back to the cloudbuilders repo.

Patch-set: 1