Let OFFLINE=True work when installing pip
Closes-Bug: #1241923

When OFFLINE=True, we should check if the system has the right pip.
If the pip is of the right version, the install pip sh exit.
If the pip is not of the right version and OFFLINE is True, the sh
complains and exits.
If the system has no pip installed at all and OFFLINE is True, the sh
complains and exits.

Change-Id: I094e6ded29d0053e9b97cb82791c4bee56d095b9