Update patch set 2
Patch Set 2: (1 inline comment)

Looks good to me except the Partial-Bug suggestion.  And while I'm nitpicking (which is a bad habit of mine ;-) it would have been a lot easier to review if separated out into multiple smaller commits such as "fix capitalization of DevStack", "clarify non-POSIX nature of DevStack", "change localrc to local.conf", and so on.  But I'm not sure it's worth the effort now to go back and break it up - just something worth bearing in mind for the future.  Thanks a lot!

Patch-set: 2
Label: Code-Review=0