Allow starting nova-compute manually
This breaks out the code that starts nova-compute into a separate
function and adds a parameter to start_nova that lets us skip
starting it. This will be used for upgrade testing so that we can
arrange for a nova-compute running on a different version of the
code to be running alongside the rest of the stack.

Change-Id: I88687cefdac7fa4a3c45789461a95fd8d061aba6