add lib/stackforge to let us test wsme / pecan from git
wsme and pecan libraries have migrated to stackforge for
development. If we support them in devstack, we can use their git
version instead of the release version, which ensures that they
won't break the rest of OpenStack when they cut a new release.

This is similar to how oslo testing works. Long term we probably
want a more generic mechanism to handle this, but for now, this
should get us rolling, and get them gating.

Change-Id: Icf3475f433081c7c625864107d7e118e214396e1