Update patch set 1
Patch Set 1:

We should probably land a change to infra to ensure that the devstack jobs are run for pecan and wsme before we land this change. That will reduce the chance that the existing master branches of both libraries will cause trouble when this does land.

Patch-set: 1
Label: Code-Review=0