Update patch set 2
Patch Set 2: Looks good to me (core reviewer)

so, while I'm fine with this, is there a reason not to create a post-processing job as part of the OpenStack gate that would automatically push the content up? It seems like it would be better to get into that pipeline for docs generation like most of the other projects do.

Patch-set: 2
Label: Code-Review=+2
Label: Workflow=0