Update patch set 2
Patch Set 2:

https://github.com/duggan/shlint is basically about removing bashisms (which we don't want, we like bashisms, we use them a lot)

shellcheck is written in haskell, and is a more generic syntax evaluation engine. Honestly haskell itself makes it a non starter.

Neither will today provide an answer to "is this script consistently indented", which is actually the problem we were trying to solve. I did really look for alternatives for this.

Patch-set: 2