Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

given this is shell, all the variables are in a global namespace by default, so I actually think we should either explicitly use 'local' to prevent the two functions from changing the same globals, or pull out the NOVA_CONF_BOTTOM logic from the functions and put it inline above them. Because  today the interaction in changing that variable would work as written, but might be a little surprising to people what's changing what and when.

Patch-set: 2
Label: Code-Review=-1
Label: Workflow=0