Update patch set 10
Patch Set 10:

This seems OK to me. I'm not in love with passing certs via environment variables but I don't have a better suggestion at the moment so lets go for it, if need be the mechanism can be enhanced later but it's better to get the functionality in ASAP. Otherwise the code looks fine to me and I would ACK it (but see below)

I see Adam says it does not work in his tree. I haven't trying running it myself and would prefer not to take the time to debug it. Have you verified it still works as expected? If so I'll test, otherwise I'll wait for an updated version.

Patch-set: 10
Label: Code-Review=0