Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

This actually doesn't work when openrc or stackrc are sourced by scripts not in the top-level devstack directory.

For example, while the exercise scripts themselves don't use BASH_SOURCE, if openrc doesn't use it then stackrc can not be found.

openrc, stackrc and eucarc are sourced in a number of places and hard-coding the relative location is not an option.  This is one reason we specifically say that we only support Bash 3.x+.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0