Update patch set 2
Patch Set 2:

I'm the one who added SWIFT_EXTRAS_MIDDLEWARE_LAST, which is not the best name ever. SWIFT_EXTRAS_MIDDLEWARE_NO_AUTH looks like a very good name to me: it makes it quite clear that it should be used for middlewares that must be added to the pipeline before the authentication modules. I'd like to keep it that way.

Patch-set: 2