Update patch set 2
Patch Set 2:

I'm not going to argue semantics, if you think that is the best name for its function, ok.  But you do use it to initialize $swift_pipeline and then only append to it, so functionally, it is the first user-configurable value even if it isn't the first value in the end pipeline.

The quote comment stands however.

Patch-set: 2