Add driver_cert wrapper for cinder
This adds a simple wrapper to call tempest volume tests.
The idea is to make it easy to execute and capture results
from tempest.api.volume.test_*

Concept is for drivers in Cinder to configure cinder.conf as
needed and then run this script which will restart services and
kick off the tempest tests, and capture the output to a logfile
for submission.

To run,
1. deploy devstack as normal with tempest included in enabled_services
2. modify cinder.conf appropriately for your driver
3. execute the script devstack/driver_certs/cinder_driver_cert.sh

Script *will* if you're not running from devstack TOP_DIR

Change-Id: I98ec9e1e418a8416406db5e2e6ffd21992e392cf