Update patch set 2
Patch Set 2:

Patch Set 2 works fine on my setup Ubuntu 12.04 on Virtualbox.
  
To be able to fix this I need an environment there I could reproduce the issues in Jenkins. 

I have reviewed the installation scripts and the fault logs and I can not find the reason why the Python tests fails in Jenkins and same tests are successful in my setup. 

The way forward now? Maybe should I "Close" this búg!
The bug is ugly but we can live with it.

Patch-set: 2
Label: Code-Review=0