Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

first, bash8 is going to fail you on this.

secondly, can we also annotate the reason for this in a comment in the code. As I think this will end up non obvious to people.

After that, I'm +2.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0