Bump node and conn timeouts, fix PKI token logging
Bump the node and connection timeouts for the swift proxy server so
that on devstack environments with constrained resources we don't get
as many timeouts for successful operations.

We also log only the first 12 characters of auth-token for the Swift
API, since PKI based auth-tokens from keystone can huge (>> 2K).

Also tidy up a comment.

Change-Id: Ib784e8ecdcb7e371fe03458c7fd82b4460fa82b9
Related-Bug: #1252514
Signed-off-by: Peter Portante <peter.portante@redhat.com>