Update patch set 2
Patch Set 2:

I really dislike this solution since it implies that keystone should be run on the wrong ports (not IANA assigned).  That being said, I don't see much else that can be done as a solution here, short of making the d-g instances use a separate port OR fix the images to use a different ephemeral port range.

I wont -1 because this is needed, but I know a lot of people use devstack as a base example deployment (to design their deployments even if this is incorrect) and this changes the default-ish behavior we use in our "public" example.

Patch-set: 2