Update patch set 2
Patch Set 2:

i'm not opposed to this merging this under one of the following conditions- either this represents openstack's complete abandonment of the IANA port assignment, or we immediately open a critical bug against devstack to resume using 35357.

my understanding is that this is merely a rare transient (although this bug # doesn't even appear on the /rechecks/ page?), so it seems apt that we should agree on a permanent solution sooner, rather than later.

Patch-set: 2