Update patch set 1
Patch Set 1:

Oh, now I see what you mean - so you have issues with including the devstack status.

So in XenAPI world, this VM is only there to run devstack, so there is a 1:1 relation in a devstack run, and this VM's life, that's why I thought that this info should be there.

Patch-set: 1