Update patch set 2
Patch Set 2:

I think /27 will be enough as it will give 32 IP addresses.
A further shortage might happen if we then increase parallelism.

I chose /27 for laziness. This way I did not have to change the gateway IP as well.
If there are no arguments against using a /24 for neutron public network I am happy to use a /24

Patch-set: 2