Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

The issue is actually slightly more complex than indicated here or in the LP bug report.

It's not just that some other process is using port 35357 as its ephemeral (automatically assigned local) port, but also that in the log output linked from the bug report (http://logs.openstack.org/09/57509/2/gate/gate-tempest-devstack-vm-large-ops/1171354/logs/screen-key.txt.gz without ?level=TRACE to see the configuration details) keystone is configured to bind to 0.0.0.0 (wildcard, any address), which is seen as * in ss/netstat output; i.e., it will accept connections destined to any local address, so long as the port matches.

The problem is that even when configured to allow multiple bindings to unique source IPs for the same port (more below), their binding will fail with EADDRINUSE (Error: Address in use) if there's already a 0.0.0.0 binding for the same local port, and vice-versa; if other non-wildcard sockets are bound to a local port, attempting to bind a wildcard to the same port will fail like in the bug log output. However, ephemeral (client) ports will *never* be sourced from 0.0.0.0, and are uniquely identified by the full connection five-tuple (proto, src IP, src port, dst IP, dst port) anyway, allowing them to overlap src IP/src port as long as proto/dst IP/dst port are different. Thus it is up to keystone/devstack to bind more appropriately and not use wildcard bind addresses unless explicitly necessary for some reason. For example, in the log output, the URLs are configured with dst IPs of 127.0.0.1 anyway, so binding explicitly to localhost would change nothing, while skirting this particular edge case nicely.

That said, you still need to set the SO_REUSEADDR sockopt to enable multiple bindings to unique source addresses for any given port. I wasn't able to determine if keystone or its socket libraries is setting this option with a cursory inspection, but if not, that would be an important step towards the right solution. I suspect it is being set as I haven't seen a keystone restart failing with EADDRINUSE anytime recently, but I also haven't particularly checked for that.

Patch-set: 2
Label: Code-Review=-1