Update patch set 1
Patch Set 1:

There's already a patch for this in glance. However, since the default stores will be FS and HTTP we need devstack to enable Swift's as well, which is required in the gates, hence this patch.

I used Partially-Fixes in the commit message since this patch doesn't fix the issue. I'll add this comment to the commit message as well.

The Glance change is: I82073352641d3eb2ab3d6e9a6b64afc99a30dcc7

Patch-set: 1
Label: Code-Review=0