Update patch set 3
Patch Set 3:

HACKING.rst is about DevStack code itself, not about using it.

I didn't say this was the wrong place, just that since you are re-working the sample file it should also be converted to the currently preferred format.  It consists of adding a single line at the top and a rename.

Commenting all of the possible values into localrc/local.conf is an exercise in futility.  You are welcome to maintain that, we will not.

Patch-set: 3