Update patch set 1
Patch Set 1:

Sean, I'd prefer removing them and relying on local.conf to set them; that is a great example of how that should work.  The only hitch is if Tempest's default values are acceptable.  If so local.conf should be the way forward.  Otherwise the name needs to be improved.

Patch-set: 1