Update patch set 1
Patch Set 1:

Chmouel, usually we run some commands, and we may probably grep the output for "fail/error/warning/denied" for something suspicious.

It would be great that we can avoid the false alarm of "/usr/sbin/apachectl: 87: ulimit: error setting limit (Operation not permitted)". Right?

If current fix is not the right way, better solution is warmly welcome. If most of us are comfortable for this error message, I'm also OK. :)

Patch-set: 1