Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Sukhdev, these are the steps to do in your workspace:

  git checkout master
  git remote update
  git pull origin master
  git checkout bug/1221781
  git rebase master
  (manually resolve the conflict in lib/neutron_plugins/ml2)
  git add lib/neutron_plugins/ml2
  git rebase --continue
  git commit --amend
  git review

For the review I am giving -1 for now since I suspect that jsonrpclib may be in the same boat as the ncclient package. Search for ncclient (git grep -w ncclient) in devstack and neutron to see how it is handled. If I am right you will need to open a bug to remove the requirement from neutron.

Patch-set: 2
Label: Code-Review=-1