Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

I am learning as I go, so please shoot down any wrongs that I write here.

As Dean pointed out, install_package is for things managed by apt (Ubuntu), yum (Redhat), etc. And jsonrpclib is not in any of those package distributions, so this is not the right fix for the bug.

Now, jsonrpclib is in pypi so it could be installed by pip_install, but pip_install is not meant for installing packages in neutron's requirements.txt. That should be done by neutron's installer.

And when neutron is installed, jsonrpclib does get installed. So the bug is something to do with the ML2 plugin somehow not seeing/loading jsonrpclib. The ML2 plugin only needs jsonrpclib if the Arista mechanism driver for ML2 is loaded.

There is a note in the bug that specifying 'name_order=True' to the ML2 type driver extension manager initializatioln fixes the issue, but I have not the foggiest clue why/how the type driver initialization would fix this, since it is Arista's mechanism driver that uses jsonrpclib.

Patch-set: 3
Reviewer: Gerrit User 6524 <6524@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1