Update patch set 3
Patch Set 3:

Hi Henry, 

Thanks for your comments, suggestions and insight. 

I wanted to provide some background and some info on some poking around that I have been doing. 
First the background: 
Since Arista is the first one to use jsonrpclib, we had bundled it with Arista driver (this is back in April/May time frame - during Grizzly release cycle). Folks on review board suggested that since this is a generic python library and would eventually be used by many others as OpenStack community grows, this should be made as a generic OpenStack requirement. Hence, it made it into the openstack/requirements.txt in Grizzly release. 

Now, back to the issue at hand. I have been using both pip install  as well apt-get install and it has been working fine. As to the best way to solve it - considering that this is a python library, I am tempted to change the code to pip install (instead of instal_package), but, am concerned it may still have some issue. I could tie it to the Arista driver, but, this will defeat the purpose and the argument made in the past. Who ever needs to use it in the future, will have to replicate the code. 

So, I am turning back to the experts on this forum to seek suggestions. 

-Sukhdev

Patch-set: 3
Reviewer: Gerrit User 6558 <6558@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0