Change ceilometer sql pool size
Because when we are running a all-in-one installation in devstack,
and ceilometer share the database with other compomnents.

To make sure that ceilometer have the time to consume all the
notifications that come from every componements during a tempest run.
We must change the default pool size to ensure it can open enough
database connection to write all the received notifications in it.

Fixes bug #1243251

Change-Id: I953d16d71a356bd7606f1b1640e7991b418c3286