Update patch set 2
Patch Set 2:

I've removed the default setting - we have a patch for neutron in review to bring it up to 10 seconds.
The current patch configures the timeout if specified.

As it stands it's an "optional" patch. One that it does not really matter if it's merged or not.

Patch-set: 2
Reviewer: Gerrit User 261 <261@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0