Add support for Fedora 20
The list of RPM packages have been updated to support the recently
released Fedora 20 distribution.

A workaround was added to the tools/fixup_stuff.sh script in order to
deal with an old version of python-crypto and the newer version that
is installed by pip.

Closes-Bug: #1263291

Co-Authored: Alvaro Lopez Ortega <alvaro@gnu.org>
Change-Id: Ia66abef1a1a54e6d5ee6eebc12908cef3f1d211d