Add libxslt and libxml2 deps to general for Fedora
Oslo incubator installation requires libxslt to be installed.  On
basic Fedora images (like the Fedora Cloud qcow2) libxslt is not
installed by default.  Running devstack on Fedora 20 fails because of
this.  The rpms/general file only specifies that libxslt and libxml2
should be installed for rhel6 distros.  This patch removes that
restriction so that those packages are installed on Fedora as well.

Change-Id: Iecf1272685a64bead5f9b2e6d2111800b731b596