Update patch set 4
Patch Set 4:

so I think the actual fix for this is adding python-lxml to rpms/cinder 

It needs >= 2.3, which should be provided fine by fc18 - fc20 by my quick googling. Then it will be pulling it from the base distro instead of compiling it from.

I suspect the fc20 fix - https://review.openstack.org/#/c/63647/ might actually address it as well. That's currently got a -1 on it because we really need to drop the unsupported fedora versions in that patch as well.

Patch-set: 4