Update patch set 2
Patch Set 2:

I left a comment on the bug, but for the sake of completeness I'll mention it here too: I wonder if this is the right behavior.  Pretty much everywhere else that names are used in OpenStack they can be duplicated, but then you are required to use the uuid to deal with those items.  I'd think this should work the same way.

Patch-set: 2
Reviewer: Gerrit User 6928 <6928@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0