Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

so I'm not really convinced we should be plumbing devstack for ceph at this point. If the extras.d set of hooks is not sufficient, we should extend that to be able to plug in ceph via an extras script.

Patch-set: 3