Update patch set 1
Patch Set 1:

Sean, thanks for the review.

I considered putting the dependency in files/, but opted for this other approach in the end. The reason is that this package needs to be installed only if the CEILOMETER_BACKEND is set to mongo. I looked for other examples of conditionally installing packages and found this:

https://github.com/openstack-dev/devstack/blob/master/lib/keystone#L387

I thought the logic here is pretty much the same (conditionally install a package, depending on the chosen backend), so I copied this approach.

Patch-set: 1
Label: Code-Review=0