Begin is_service_enabled() cleanup
This converts the special cases in the is_service_enabled() function to call
individual functions declared by the projects.  This allows projects that
are not in the DevStack repo and called via the extras.d plugin to handle
an equivalent service alias.

* Ceilometer
* Cinder
* Glance
* Neutron
* Nova
* Swift

Change-Id: Ic78be433f93a9dd5f46be548bdbd4c984e0da6e7
TODO: remove the tests from is_service_enabled() after a transition period