Update patch set 1
Patch Set 1:

Because of the boto issue we pulled the bundle and euca exercises from the default list, should we just do that here too?

I guess the greater question is how much attention do we need to pay to the EC2 API at this point?

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0