Set default adapter type for monolithic flat disks
upload_image.sh tries to find corresponding metadata in the
descriptor file for a monolithicFlat *-flat.vmdk
If this descriptor is not found, the script sets the disk_type
to preallocated but doesn't set the adapter_type.
Most of the VMs will fail booting without an adapter type.
This patch sets the adapter type to a default value: this value might
not the correct/most optimized but at least it a value is set.

Booting the example VMDK provided
https://wiki.openstack.org/wiki/NovaVMware/DeveloperGuide will work.

Closes-Bug: #1271793

Change-Id: Ie3c23a4bb212f30fd078cec6b2b1d604738b595d