Update patch set 2
Patch Set 2:

Sean, the tempest config parser is looking for "TEMPEST_CONFIG" so the existing variable names in devstack don't actually work.  Maybe there's something I'm missing but when I tried to set this up it didn't work so this is what I had to do, it seemed to me that there was a change in one of the other at some point that didn't synch.

This naming of tempest config is actually the primary motivation/need for the patch, I'll make that more obvious in the commit messages.

vendor_name and protocol are BOTH used in the type test when setting extra_specs.  These are key values for the cinder capabilities filter, so if you use them in the extra-specs they MUST be available or you get "No valid host available" when it fails the capability filter checks.

Patch-set: 2