Update patch set 8
Patch Set 8: I would prefer that you didn't merge this

(11 inline comments)

I would like to see the backend plugin stuff look more like the hypervisor plugins for nova, this is really similar to the requirements that Docker has: an external service that needs to be installed from a 3rd party repo and config changes to existing services need to be made.

At this point I'm not ready to include the ceph files themselves in the repo but we can make the changes to support dropping them in place and having everything work.

Patch-set: 8
Label: Code-Review=-1
Label: Workflow=0