Update patch set 2
Patch Set 2:

The docker driver for Nova needs a registry service to be running. It is being run inside a container using an image -- that image must be downloaded.  The registry service must be started via nova_plugins/hypervisor-docker, but this is presently called before Glance's image download.

The reordering is being done such that Glance may download the registry image, but prior to starting Nova such that "hypervisor-docker" may have an image downloaded and available to launch the registry.

Patch-set: 2