Add support for the OpenDaylight Neutron MechanismDriver
This commit adds support for running OpenDaylight under devstack. Changes are
also added to configure the ML2 OpenDaylight MechanismDriver in both the case
of running OpenDaylight under devstack, as well as in the case where you run
OpenDaylight externally to devstack, as is likely the case for OpenDaylight
developers wanting to test with devstack.

Closes-bug: #1273917

Change-Id: I696e7c7fe63c835f90c56105775def305a702877