Add support for configuring OVS to work with OpenDaylight
This adds support for enabling OpenDaylight in devstack. This first
cut simply allows you to "enable_service opendaylight", in which case
this will trigger the configuration of OVS on the host to point both
it's OVSDB and per-bridge OpenFlow connections at the OpenDaylight
controller.

Future patches will enable OpenDaylight to be run under devstack.

Closes-bug: #1273917

Change-Id: I696e7c7fe63c835f90c56105775def305a702877