Update patch set 10
Patch Set 10: I would prefer that you didn't merge this

You are not in the minority John, I do not intend to add Ceph directly at this time, but I will add the hooks necessary to use it similar to how the hypervisors are handled in nova.  Then using Ceph would be simply adding the specific files to a devstack tree and it just works.

Patch-set: 10
Label: Code-Review=-1
Label: Workflow=0