Update patch set 2
Patch Set 2:

Phew, that was fun.  I made the requested changes and found a new way this can fail.  If you log in to the instance before cloud-init has set the hostname, rabbitmqctl tries to connect to the wrong hostname and fails until you log back in.

It might be possible to fix that by specifying the node as rabbit@`hostname -s` in the calls to rabbitmqctl, but I'm unsure what other side-effects that could have so I didn't do it in this patch set.

Patch-set: 2
Reviewer: Gerrit User 6928 <6928@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0