Install libguestfs for nova-compute on Ubuntu
We were already installing this for n-cpu on rpm distros, but not
Ubuntu.  Install it so that nova-compute can use it for file injection,
which is the preferred method over nbd.

Add a check for libguestfs and set CONF.libvirt.inject_partition to -1 if
it is available.  This enables using libguestfs to determine the proper
partition to inject into.

If libguestfs is available, don't bother trying to load the nbd kernel module,
as it won't be used.

Change-Id: Ifa9d95bf759f1dad8685590a2df242d852dd2cb0