Install libguestfs for nova-compute on Ubuntu
We were already installing this for n-cpu on rpm distros, but not
Ubuntu.  Install it so that nova-compute can use it for file injection,
which is the preferred method over nbd.

(Pulled out of gate, this actually didn't pass check, and turns out
guestfs is actually broken in icehouse.... yay.... tests?)

Change-Id: Ifa9d95bf759f1dad8685590a2df242d852dd2cb0