Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(3 inline comments)

It is a corner case but this changes the behaviour in a subtle way in multi-node installations.  These are operations that should only happen once for a given cluster and they test for a service that should only be running once in a given cluster to achieve that.  Using the generic service name will attempt to do this on every node that runs any of those services.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0