Update patch set 3
Patch Set 3:

Some functionality of lib/ironic is relying on lib/baremetal, and some is duplicating similar methods. 

Ultimately, nova "baremetal" driver will be deprecated, and so devstack/lib/baremetal will be removed at that time. We should have a clear path towards that. Either:

a) put ironic support also in lib/baremetal, make it support one-or-the-other with an ENV var, and then we can remove the baremetal methods later

b) keep the files separate, and refactor any common methods into a shared lib, so that we can simply delete lib/baremetal when that time comes.

Patch-set: 3