read_password needs to store in .localrc.auto if local.conf is used.
With local.conf, When we are asking to complete a mandatory field,
read_password create a file localrc to store it. This situation
adds a problem when stackrc try to source env variables.

To fix this situation when we use the new local.conf, we need to
store options in .localrc.auto.

https://github.com/openstack-dev/devstack/blob/master/stackrc#L60

Related-Bug #1260723
Change-Id: I25ad07569d2b42b190449591d5a01ade8022392c