Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(1 inline comment)

In order to sort this out in my mind I want to list a couple of givens:

* .localrc.auto is removed near the top of stack.sh on every run, so the persistence of these passwords is now lost if using local.conf.  I consider this behaviour intentional, persistent passwords should be in local.conf.
* local.conf should NEVER be modified by any DevStack component.  This differs from the localrc case with user-entered passwords.  I never liked modifying user config files.

Given that, the potential problem I see is when running stack.sh with no passwords in local.conf, read_password() creates localrc and local.conf is ignored on subsequent runs of stack.sh.

The commit message implies a situation that is not clear to me...by the time read_password() is used stackrc has already sourced everything it is going to and the results of read_password() will be available to the rest of the script.

Is there another use case missing here?

Patch-set: 3
Label: Code-Review=-1