Update patch set 4
Patch Set 4:

Sean: The reason it's done this way is to allow for both the case where OpenDaylight runs under devstack, as well as the case where OpenDaylight is run externally to devstack, which would be the case for OpenDaylight developers. Splitting it so configuration is handled this way allows both things to work at the same time.

Patch-set: 4
Label: Code-Review=0