Create glance-swift for swift or s-proxy
The keystone_data.sh file creates the glance-swift user which is
required to use glance with swift - however it only does it when
the s-proxy service is enabled. To be consistent with other parts
of the script, if should also create the user if the swift service
is enabled. In any case something appears to be dropping s-proxy
from the list of enabled services; despite explicitly enabling
both swift and s-proxy in my localrc, only swift is seen.

Closes-bug : #1279323
Change-Id: Ibec9f03755bd71e8736a3a81bc668e6f8f103c9a