Update patch set 9
Patch Set 9: I would prefer that you didn't merge this

We should not be adding tripleo-incubator to the check/gate pipeline for Ironic (and thus eventually to many other projects) because tripleo-incubator doesn't guarantee any API compatibility and is likely to change. 

I initially suggested this -- what I really meant was, we should copy the functionality we need and modify / maintain it here.

Patch-set: 9
Label: Code-Review=-1