fix bash8 failure condition
bash8 wasn't reseting the multiline variable at the end of a run,
so it was actually failing to catch errors in a file if it was
tricked into multiline at the end of a previous one.

redo how we are processing the file list so this is reset before
each file starts.

Change-Id: Ic85d884088d8a84e6f34f2b642559c1017fe1378