Update patch set 6
Patch Set 6: I would prefer that you didn't merge this

so I think we've decided that setting up and installing the SDN is outsite of the scope of devstack, unless it's as simple as a package install included in the distro. So I think we should remove that piece of this.

Patch-set: 6
Label: Code-Review=-1
Label: Workflow=0