Update patch set 7
Patch Set 7:

I think having this in extras is the way to go here. What I'm doing is configuring OVS on the host (and tearing it down) so it's setup to work with OpenDaylight. I don't see the right hooks in the lib/neutron_plugins/$Q_PLUGIN infra for this, and in fact I thought this is what extras was designed for.

Second, I see existing Neutron plugins (like Ryu) which are doing "git pulls" and running SDN controllers in devstack. So I feel as if perhaps this was grandfathered in, but nonetheless, it's an inconsistent policy at best.

Third, this is a pure Open Source controller, and the goal is to get this running in the gate eventually.

Please reconsider your -1 here and/or ping me on IRC to discuss, I'll look for you as well.

Patch-set: 7
Label: Code-Review=0