Update patch set 11
Patch Set 11: I would prefer that you didn't merge this

(5 inline comments)

Does Ironic support more than Ubuntu?  If so the prereq files need to be present. If not, there should be some checks to gracefully abort on un-supported platforms.

The scripts in tools/ironic/scripts need to have some semi-reasonable error handling.  I'll not hold up this review for that, but please follow up with it soon.

Patch-set: 11
Label: Code-Review=-1
Label: Workflow=0